refactor: clarify Reset button text and address PR feedback
- Extracted ResetSection component to standalone file at `app/client/connect/components/ResetSection.tsx`.
- Updated button text to 'Reset Permissions & Settings' and clarified description.
- Refactored styling to use MUI system properties (`borderTop: 1`, `borderColor: 'divider'`).
- Renamed VRT snapshot in test to 'connect-page-scanning.png' to match project history and satisfy reviewer feedback.
- Updated all unit and integration tests.
- Regenerated all VRT snapshots.
