+++ tests/playwright/vrt-connect-page.spec.ts-snapshots/connect-page-scanning-chromium-linux.png
